requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
asn1crypto==0.24.0
asttokens==2.0.5
async-timeout==4.0.2
asynctest==0.13.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bidict==0.21.4
certifi==2021.10.8
cffi==1.14.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==2.4.2
dateparser==1.1.0
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
flask-sock==0.5.0
Flask-SocketIO==5.1.1
frozenlist==1.3.0
h11==0.13.0
hyperlink==18.0.0
idna==2.8
importlib-metadata==4.11.2
incremental==17.5.0
ipdb==0.13.9
ipython==7.32.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
kraken-wsclient-py==0.0.9
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
Pygments==2.11.2
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-binance==1.0.15
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
service-identity==18.1.0
simple-websocket==0.5.1
six==1.12.0
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
Twisted==19.7.0
txaio==18.8.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.8
wcwidth==0.2.5
websockets==9.1
Werkzeug==2.0.3
wsproto==1.1.0
yarl==1.7.2
zipp==3.7.0
zope.interface==4.6.0
